1 Add card partial
Card partial contains the card that is used to display the blog
information.The card.css has styles that are very specific to the card
and the additions to the index.css are are styles that will most likely
be used again. Please note that, general styles will be applied using
utility classes, thus the very few rules in the classes added to index.

Last change is adding font-size of 16px to html as that will be the base
of relative sizes in the blog.